feat: Validation is described in a separate module
